Revert "Merge branch 'refactor/407' into ros2launchman"
This reverts commit 6b985e0a5248de0a9e38f8f1a2825477405a58f3, reversing
changes made to 5868d6686f83b78702c0a76a8d746df4ea3e3046.
