+++ src/roswire/proxy/__init__.py
@@ -2,4 +2,5 @@
 from .bag import BagPlayer, BagRecorder
 from .catkin import CatkinInterface, CatkinMake, CatkinTools
 from .roscore import ROSCore
+from .roslaunch import ROSLaunchManager
 from .state import SystemState
