+++ src/roswire/proxy/roslaunch/__init__.py
@@ -1,1 +1,2 @@
 # -*- coding: utf-8 -*-
+from .roslaunch import ROSLaunchManager
