+++ src/roswire/ros1/__init__.py
@@ -1,5 +1,3 @@
 # -*- coding: utf-8 -*-
 from .manager import ROS1NodeManager
 from .node import ROS1Node
-from .reader import ROS1LaunchFileReader
-from .roslaunch import ROS1LaunchManager
