+++ src/roswire/ros2/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-from .manager import ROS2NodeManager
+from .node_manager import ROS2NodeManager
 from .ros2 import ROS2
 from .service_manager import ROS2ServiceManager
