+++ src/roswire/ros2/ros2.py
@@ -6,7 +6,7 @@ import typing
 import attr
 
 from .launch import ROS2LaunchManager
-from .manager import ROS2NodeManager
+from .node_manager import ROS2NodeManager
 from .service_manager import ROS2ServiceManager
 from .state import ROS2StateProbe
 from ..proxy import SystemState
